댓글 수정 버그 수정, 댓글 입력 글자 카운트 버그 수정

+++ main/java/com/models/board/BoardDao.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletRequest;
 
 import com.core.*;
 import com.models.board.*;
+import com.models.file.*;
 import com.models.member.*;
 
 public class BoardDao {
@@ -28,26 +29,30 @@ public class BoardDao {
 
 		/** 유효성 검사 **/
 		checkData(request);
-
+		
+		Member member = (Member)request.getAttribute("member");
+		
 		int num = 0;
-		String sql = "INSERT INTO board (status, postTitle, content, memId, isNotice) VALUES(?,?,?,?,?)";
+		String sql = "INSERT INTO board (gid, status, postTitle, content, memId, isNotice) VALUES(?,?,?,?,?,?)";
 		try (Connection conn = DB.getConnection();
 				PreparedStatement pstmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 			request.setCharacterEncoding("UTF-8");
+			String gid = request.getParameter("gid");
 			String status = request.getParameter("status");
 			String postTitle = request.getParameter("postTitle");
 			String content = request.getParameter("content");
-			String memId = request.getParameter("memId");
+			String memId = member.getMemId();
 			int isNotice = 0;
 			if (request.getParameter("isNotice") != null) {
 				isNotice = Integer.valueOf(request.getParameter("isNotice"));
 			}
 
-			pstmt.setString(1, status);
-			pstmt.setString(2, postTitle);
-			pstmt.setString(3, content);
-			pstmt.setString(4, memId);
-			pstmt.setInt(5, isNotice);
+			pstmt.setLong(1, Long.valueOf(gid));
+			pstmt.setString(2, status);
+			pstmt.setString(3, postTitle);
+			pstmt.setString(4, content);
+			pstmt.setString(5, memId);
+			pstmt.setInt(6, isNotice);
 
 			int result = pstmt.executeUpdate();
 			if (result > 0) {
@@ -56,6 +61,8 @@ public class BoardDao {
 					num = rs.getInt(1);
 				}
 				rs.close();
+				
+				FileDao.getInstance().updateFinish(gid);
 			}
 		} catch (IOException | SQLException | ClassNotFoundException e) {
 			Logger.log(e);
@@ -166,6 +173,7 @@ public class BoardDao {
 
 			int rs = pstmt.executeUpdate();
 			if (rs > 0) {
+				FileDao.getInstance().updateFinish(request.getParameter("gid"));
 				return true;
 			}
 		} catch (Exception e) {
@@ -175,32 +183,85 @@ public class BoardDao {
 		return false;
 	}
 
-	public boolean delete(int postNm) {
-
+	public boolean delete(int postNm) throws Exception {
+		
+		/**
+		 *  1. 로그인 여부
+		 *  2. 게시글 조회 -> 게시글 존재 유부, 게시글에 등록된 아이디와 로그인한 아이들 비교
+		 *  3. 일치하는 경우만 삭제(본인 게시글만 삭제 가능) 
+		 */ 
+		HttpServletRequest request = Req.get();
+		if (!MemberDao.isLogin(request)) {
+			throw new Exception("로그인 후 삭제가 가능합니다."); 
+		}
+		
+		Board board = get(postNm);
+		if (board == null) {
+			throw new Exception("존재하지 않는 게시글 입니다.");
+		}
+		
+		Member member = (Member)request.getAttribute("member");
+		if (!member.getMemId().equals(board.getMemId())) {
+			throw new Exception("본인이 작성한 게시글만 삭제할 수 있습니다.");
+		}
+		
 		String sql = "DELETE FROM board WHERE postNm=?";
-		try (Connection conn = DB.getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
-			pstmt.setInt(1, postNm);
-
-			int rs = pstmt.executeUpdate();
-			if (rs > 0) {
-				return true;
-			}
-
-		} catch (Exception e) {
-			Logger.log(e);
+		ArrayList<DBField> bindings = new ArrayList<>();
+		bindings.add(DB.setBinding("Integer", String.valueOf(postNm)));
+		int rs = DB.executeUpdate(sql, bindings);
+		
+		/** 게시글 DB 삭제 후 업로드된 이미지 파일 삭제 */
+		if (rs > 0) {
+			FileDao.getInstance().deleteByGid(board.getGid());
 		}
-		return false;
+		
+		return (rs > 0)?true:false;
 	}
-
+	
+	public boolean delete(String postNm) throws Exception {
+		return delete(Integer.valueOf(postNm));
+	}
+	
+	
 	public void checkData(HttpServletRequest req) throws Exception {
-
-		String[] required = { "status//어떤게시글인지 정해주세요.", "postTitle//글 제목을 입력해주세요.", "content//내용을 입력해주세요." };
+		if (!MemberDao.isLogin(req)) {
+			throw new Exception("로그인이 필요한 서비스입니다.");
+		}
+		
+		String[] required = {
+				"gid//잘못된 접근입니다",
+				"status//어떤게시글인지 정해주세요.", 
+				"postTitle//글 제목을 입력해주세요.", 
+				"content//내용을 입력해주세요." 
+			};
 		for (String s : required) {
 			String[] re = s.split("//");
 			if (req.getParameter(re[0]) == null || req.getParameter(re[0]).trim().equals("")) {
 				throw new Exception(re[1]);
 			}
 		}
+	}
+	
+	/**
+	 * 게시글 조회수 업데이트
+	 * 
+	 * @param postNm
+	 */
+	public void updateViewCnt(int postNm) {
+		String browserId = CommonLib.getBrowserId();
+		try {
+			String sql = "INSERT INTO boardview VALUES (?, ?)";
+			ArrayList<DBField> bindings = new ArrayList<>();
+			bindings.add(DB.setBinding("Integer", String.valueOf(postNm)));
+			bindings.add(DB.setBinding("String", String.valueOf(browserId)));
+			
+			DB.executeUpdate(sql, bindings);
+			
+			sql = "UPDATE board a SET a.viewCnt = (SELECT COUNT(*) FROM boardview b WHERE a.postNm = b.postNm) WHERE a.postNm = ?";
+			bindings = new ArrayList<DBField>();
+			bindings.add(DB.setBinding("Integer", String.valueOf(postNm)));
+			DB.executeUpdate(sql, bindings);
+			
+		} catch (Exception e) {}
 	}
 }
